requirements.txt
aiohttp==3.4.2
async-timeout==3.0.0
attrs==18.2.0
chardet==3.0.4
idna==2.7
idna-ssl==1.1.0
marshmallow==2.15.4
multidict==4.3.1
pexpect==4.6.0
python-dateutil==2.7.3
six==1.11.0
uvloop==0.11.2
yarl==1.2.6
